rename internal to track builder outside of formats in preparation for api redesign
