+++ src/formats.rs
@@ -1,4 +1,3 @@
-pub mod internal;
 pub mod lrb;
 pub mod sol;
 pub mod trackjson;
