+++ src/track_builder/track_line/scenery.rs
@@ -1,4 +1,4 @@
-use crate::internal::track_line::Line;
+use crate::track_builder::track_line::Line;
 
 #[derive(Debug, Clone)]
 pub struct SceneryLine {
