+++ src/track_builder/trigger/background_color.rs
@@ -1,7 +1,5 @@
-use crate::internal::{rgb_color::RGBColor, trigger::FrameReachedEvent};
-// TODO: Make this per layer
+use crate::track_builder::{rgb_color::RGBColor, trigger::FrameReachedEvent};
 
-// TODO: This is intended for LRO color triggers acting on the base layer
 #[derive(Debug, Clone)]
 pub struct State {
     pub color: RGBColor,
